+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 -   The search popover is opened if users press the shortcuts (`Ctrl + F`, or `Cmd + F` on macOS) when the mouse is inside the viewer container
 -   It's able to scroll the pages when opening the search popover
+-   Support link annotations that have the `unsafeUrl` property
 
 **Bug fixes**